trex_runner/requirements.txt
numpy==1.21.1
mss==6.1.0
Pillow==8.3.1
opencv-python==4.5.3.56
keyboard==0.13.5
